+++ dev/20190415/k2works/fizzbuzz/src/test/java/org/hiroshima_arc/presenntaion/controller/StreamLambdaHandlerTest.java
@@ -1,4 +1,4 @@
-package org.hiroshima_arc;
+package org.hiroshima_arc.presenntaion.controller;
 
 
 import com.amazonaws.serverless.proxy.internal.LambdaContainerHandler;
@@ -7,6 +7,7 @@ import com.amazonaws.serverless.proxy.internal.testutils.MockLambdaContext;
 import com.amazonaws.serverless.proxy.model.AwsProxyResponse;
 import com.amazonaws.services.lambda.runtime.Context;
 
+import org.hiroshima_arc.StreamLambdaHandler;
 import org.junit.BeforeClass;
 import org.junit.Test;
